resolve vi missing dependency
By design vitest does not allow mocking of a function called by a function. A better approach, as used here is to inject the dependency in when testing. A slight modification to thr Team.ts domain to use the default team.dal.ts and then creation of a new MockedDataAccessLayer for testing purposes allows for this.